+++ test/test_helper.rb
@@ -15,7 +15,7 @@ ENV['NO_PROXY'] ||= ENV['UNICORN_TEST_ADDR'] || '127.0.0.1'
 # Some tests watch a log file or a pid file to spring up to check state
 # Can't rely on inotify on non-Linux and logging to a pipe makes things
 # more complicated
-DEFAULT_TRIES = 1000
+DEFAULT_TRIES = 100
 DEFAULT_RES = 0.2
 
 require 'minitest/autorun'